+++ Lusas_Adapter/Convert/ToBHoM/Loads/ToPointDisplacement.cs
@@ -63,7 +63,7 @@ namespace BH.Adapter.Adapters.Lusas
                 loadcase, assignedNodes, translationVector, rotationVector, LoadAxis.Global,
                 GetName(lusasPrescribedDisplacement));
 
-            int adapterNameId = (int)lusasPrescribedDisplacement.getID();
+            long adapterNameId = lusasPrescribedDisplacement.getID();
             pointDisplacement.SetAdapterId(typeof(LusasId), adapterNameId);
             // Needs to be a bit here that determines whether it is global or local - actually this cannot be done as the 
             //attribute is applied to a group, and within the group the axis could local or global
